Implement Member Given Points History

+++ pointdistribution/core/points_operation.py
@@ -1,6 +1,9 @@
 from .exceptions import RepeatedPointValueException, MembersMissingException, InvalidSumPointsException, \
     ConflictInPointsToMemberException, InvalidOrRepeatedMemberException, PointValueNotValidException
 from .models import GivenPoint
+from .serializers import GivenPointArchivedSerializer
+from .exceptions import InvalidGivenPointsArchivedData
+from django.forms.models import model_to_dict
 
 
 def validate_provisional_point_distribution(point_distribution, members_list):
@@ -19,6 +22,13 @@ def validate_provisional_point_distribution(point_distribution, members_list):
     if len(member_to_point) != len(members_list):
         raise MembersMissingException()
     for given_point in given_points:
+        given_point_dict = model_to_dict(given_point)
+        del given_point_dict['point_distribution']
+        del given_point_dict['id']
+        serializer = GivenPointArchivedSerializer(data=given_point_dict)
+        if not serializer.is_valid():
+            raise InvalidGivenPointsArchivedData()
+        serializer.save()
         given_point.delete()
     sum_points = 0
     week = point_distribution.week
